+++ src/rust/src/helpers/sparse_graph_dict.rs
@@ -147,7 +147,7 @@ pub struct DgrdlResults {
 /// * `data_hash` - Hash of the underlying data
 /// * `laplacian_cache` - HashMap of the feature and sample Laplacian matrix
 /// * `dictionary_cache` - HashMap of the dictionary cashe with (dict_size, seed)
-///                        as keys
+///   as keys
 /// * `distance_matrix` - The distance matrix of the data as a flat vector
 #[derive(Debug)]
 pub struct DgrdlCache {
@@ -229,12 +229,14 @@ impl DgrdlCache {
 /// ### Fields
 ///
 /// * `approximation_error` - Calculates the approximation error in form of
-///                           squared Frobenius norm
+///   squared Frobenius norm
 /// * `feature_laplacian_objective` - Measures the similiarity of coefficients
-///                                   for the feature Laplacian
+///   for the feature Laplacian
 /// * `sample_laplacian_objective` - Measures the smoothness of the dictionary
-///                                  in respect to sample Laplacian
-/// *
+///   in respect to sample Laplacian
+/// * `seed` - Random seed for reproducibility
+/// * `k_neighbours` - Number of neighbours
+/// * `dict_size` - The size of the dictionary
 #[derive(Debug, Clone)]
 pub struct DgrdlObjectives {
     pub approximation_error: f64,
@@ -840,10 +842,10 @@ fn get_distance(distances: &[f64], i: usize, j: usize, n: usize) -> f64 {
 /// ### Fields
 ///
 /// * `data` - The original data matrix with rows = samples and columns =
-///            features
+///   features
 /// * `k` - Number of neighbours for the KNN graph
 /// * `features` - If `true` generated the Laplacian for the features, otherwise
-///                for the samples
+///   for the samples
 ///
 /// ### Returns
 ///